refactor: Improve lock handling and code organization across runtime
Multiple improvements across the async runtime codebase:

1. net/dns/cache.rs: Reduce lock contention by cloning entries before
   releasing the read lock, preventing lock-under-lock scenarios.
   Change find_oldest_key to associated function (no &self needed).

2. trace/replay.rs: Reorganize estimated_size() match arms to group
   variants by their serialized size, improving readability.

3. Various modules: Minor style cleanups, improved error handling,
   and code organization improvements.

Files modified:
- src/net/dns/{cache,resolver}.rs
- src/net/tcp/{listener,stream,traits}.rs
- src/net/unix/ancillary.rs
- src/record/task.rs
- src/runtime/{builder,deadline_monitor,io_driver,state,stored_task}.rs
- src/service/retry.rs
- src/stream/fuse.rs
- src/test_utils.rs
- src/time/wheel.rs
- src/trace/{compat,filter,replay,replayer}.rs
- src/transport/tests.rs
- src/types/rref.rs
- src/lab/oracle/determinism.rs

+++ src/net/unix/ancillary.rs
@@ -158,6 +158,7 @@ impl<'a> SocketAncillary<'a> {
     ///     panic!("Buffer too small for file descriptors");
     /// }
     /// ```
+    #[allow(clippy::cast_ptr_alignment)]
     pub fn add_fds(&mut self, fds: &[RawFd]) -> bool {
         if fds.is_empty() {
             return true;
@@ -250,6 +251,7 @@ pub struct AncillaryMessages<'a> {
 impl<'a> Iterator for AncillaryMessages<'a> {
     type Item = AncillaryMessage<'a>;
 
+    #[allow(clippy::cast_ptr_alignment)]
     fn next(&mut self) -> Option<Self::Item> {
         if self.current >= self.buffer.len() {
             return None;
@@ -269,7 +271,7 @@ impl<'a> Iterator for AncillaryMessages<'a> {
                 return None;
             }
 
-            let cmsg_len = (*cmsg_ptr).cmsg_len as usize;
+            let cmsg_len = (*cmsg_ptr).cmsg_len;
             if cmsg_len < mem::size_of::<libc::cmsghdr>() {
                 return None;
             }
